pyproject.toml
[project]
name = "citegeist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.3",
    "networkx>=3.5",
    "numpy>=2.3.2",
    "plotly>=6.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
